refactor: VisualTransformation 분리
