+++ app/src/main/java/woowacourse/payments/ui/newcard/textfields/ExpirationDateTextField.kt
@@ -11,16 +11,19 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.OffsetMapping
-import androidx.compose.ui.text.input.TransformedText
-import androidx.compose.ui.text.input.VisualTransformation
 import woowacourse.payments.R
 import woowacourse.payments.domain.ExpirationDate
+import woowacourse.payments.ui.formatter.UniformlySeparatingVisualTransformation
 import woowacourse.payments.ui.theme.Gray
 import java.time.YearMonth
 import java.time.format.DateTimeFormatter
+
+private const val EXPIRATION_DATE_REQUIRED_LENGTH = 4
+private const val EXPIRATION_DATE_CHUNK_SIZE = 2
+private const val EXPIRATION_DATE_SEPARATOR = " - "
+private val visualTransformation =
+    UniformlySeparatingVisualTransformation(EXPIRATION_DATE_CHUNK_SIZE, EXPIRATION_DATE_SEPARATOR)
 
 @Suppress("ktlint:standard:function-naming")
 @Composable
@@ -29,36 +32,6 @@ fun ExpirationDateTextField(
     isError: MutableState<Boolean>,
 ) {
     val focusManager = LocalFocusManager.current
-    val delimiter = stringResource(R.string.expiration_date_delimiter)
-
-    val offsetMapping =
-        object : OffsetMapping {
-            override fun originalToTransformed(offset: Int): Int {
-                val multiplier = (offset - 1).coerceAtLeast(0) / EXPIRATION_DATE_CHUNK_SIZE
-                return offset + delimiter.length * multiplier
-            }
-
-            override fun transformedToOriginal(offset: Int): Int {
-                val multiplier =
-                    ((offset - 1)).coerceAtLeast(0) / (EXPIRATION_DATE_CHUNK_SIZE + delimiter.length)
-                return (offset - (delimiter.length * multiplier)).coerceAtMost(
-                    EXPIRATION_DATE_CHUNK_SIZE * (multiplier + 1),
-                )
-            }
-        }
-
-    val transformation =
-        object : VisualTransformation {
-            override fun filter(text: AnnotatedString): TransformedText =
-                TransformedText(
-                    AnnotatedString(
-                        text.text
-                            .chunked(EXPIRATION_DATE_CHUNK_SIZE)
-                            .joinToString(delimiter),
-                    ),
-                    offsetMapping,
-                )
-        }
 
     fun updateValue(newValue: String) {
         val filteredValue: String =
@@ -85,7 +58,7 @@ fun ExpirationDateTextField(
         value = text.value,
         onValueChange = { newValue: String -> updateValue(newValue) },
         singleLine = true,
-        visualTransformation = transformation,
+        visualTransformation = visualTransformation,
         label = { Text(stringResource(R.string.expiration_date_label)) },
         placeholder = {
             Text(
@@ -107,6 +80,3 @@ fun ExpirationDateTextField(
         keyboardActions = KeyboardActions(onDone = { focusManager.moveFocus(FocusDirection.Next) }),
     )
 }
-
-private const val EXPIRATION_DATE_REQUIRED_LENGTH = 4
-private const val EXPIRATION_DATE_CHUNK_SIZE = 2
